+++ src/backend/joanie/payment/tests/test_admin_invoice.py
@@ -1,4 +1,4 @@
-"""Invoice admin test suite"""
+"""ProformaInvoice admin test suite"""
 from django.test import TestCase
 from django.urls import reverse
 
@@ -6,28 +6,29 @@ import lxml.html
 from djmoney.money import Money
 
 from joanie.core import factories
-from joanie.payment.factories import InvoiceFactory, TransactionFactory
+from joanie.payment.factories import ProformaInvoiceFactory, TransactionFactory
 
 
-class InvoiceAdminTestCase(TestCase):
-    def test_admin_invoice_display_human_readable_type(self):
+class ProformaInvoiceAdminTestCase(TestCase):
+    def test_admin_proforma_invoice_display_human_readable_type(self):
         """
-        Invoice admin view should display invoice type in a human-readable manner.
+        ProformaInvoice admin view should display
+        pro forma invoice type in a human-readable manner.
         """
         # - Login as admin
         user = factories.UserFactory(is_staff=True, is_superuser=True)
         self.client.login(username=user.username, password="password")
 
-        # - Create an order with a related invoice
+        # - Create an order with a related pro forma invoice
         order = factories.OrderFactory()
-        invoice = InvoiceFactory(order=order, total=order.total)
-        credit_note = InvoiceFactory(
+        invoice = ProformaInvoiceFactory(order=order, total=order.total)
+        credit_note = ProformaInvoiceFactory(
             order=invoice.order, parent=invoice, total=-order.total
         )
 
-        # - Now go to the invoice admin change view
+        # - Now go to the pro forma invoice admin change view
         response = self.client.get(
-            reverse("admin:payment_invoice_change", args=(credit_note.pk,)),
+            reverse("admin:payment_proformainvoice_change", args=(credit_note.pk,)),
         )
 
         self.assertEqual(response.status_code, 200)
@@ -39,25 +40,25 @@ class InvoiceAdminTestCase(TestCase):
         self.assertEqual(credit_note.type, "credit_note")
         self.assertEqual(type_field.text_content(), "Credit note")
 
-    def test_admin_invoice_display_balances(self):
+    def test_admin_proforma_invoice_display_balances(self):
         """
-        Invoice admin view should display invoice balances
+        ProformaInvoice admin view should display pro forma invoice balances
         (global, transactions and invoiced).
         """
         # - Login as admin
         user = factories.UserFactory(is_staff=True, is_superuser=True)
         self.client.login(username=user.username, password="password")
 
-        # - Create an order with a related invoice
+        # - Create an order with a related pro forma invoice
         order = factories.OrderFactory()
-        invoice = InvoiceFactory(
+        invoice = ProformaInvoiceFactory(
             order=order, total=Money(order.total.amount, currency="XYZ")
         )
-        InvoiceFactory(order=invoice.order, parent=invoice, total=-order.total)
+        ProformaInvoiceFactory(order=invoice.order, parent=invoice, total=-order.total)
 
-        # - Now go to the invoice admin change view
+        # - Now go to the pro forma invoice admin change view
         response = self.client.get(
-            reverse("admin:payment_invoice_change", args=(invoice.pk,)),
+            reverse("admin:payment_proformainvoice_change", args=(invoice.pk,)),
         )
 
         self.assertEqual(response.status_code, 200)
@@ -74,34 +75,35 @@ class InvoiceAdminTestCase(TestCase):
         self.assertEqual(transactions_balance_field.text_content(), "XYZ0.00")
         self.assertEqual(invoiced_balance_field.text_content(), "XYZ0.00")
 
-    def test_admin_invoice_display_invoice_children_as_link(self):
+    def test_admin_proforma_invoice_display_invoice_children_as_link(self):
         """
-        Invoice admin view should display list of invoice children if there are.
+        ProformaInvoice admin view should display list of
+        pro forma invoice children if there are.
         """
         # - Login as admin
         user = factories.UserFactory(is_staff=True, is_superuser=True)
         self.client.login(username=user.username, password="password")
 
-        # - Create an order with a related invoice
+        # - Create an order with a related pro forma invoice
         order = factories.OrderFactory()
-        invoice = InvoiceFactory(order=order, total=order.total)
+        invoice = ProformaInvoiceFactory(order=order, total=order.total)
 
-        # - And link other invoices to this invoice
-        children = InvoiceFactory.create_batch(
+        # - And link other pro forma invoices to this pro forma invoice
+        children = ProformaInvoiceFactory.create_batch(
             2, order=invoice.order, parent=invoice, total=order.total
         )
 
-        # - Now go to the invoice admin change view
+        # - Now go to the pro forma invoice admin change view
         response = self.client.get(
-            reverse("admin:payment_invoice_change", args=(invoice.pk,)),
+            reverse("admin:payment_proformainvoice_change", args=(invoice.pk,)),
         )
 
-        # - Invoice are ordered by creation date
+        # - ProformaInvoice are ordered by creation date
 
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, invoice.reference)
 
-        # - Check there are links to go to invoice children admin change view
+        # - Check there are links to go to pro forma invoice children admin change view
         html_parser = lxml.html.HTMLParser(encoding="utf-8")
         html = lxml.html.fromstring(response.content, parser=html_parser)
         invoice_children_field = html.cssselect(".field-children")[0]
@@ -114,7 +116,7 @@ class InvoiceAdminTestCase(TestCase):
         )
         self.assertEqual(
             links[0].attrib["href"],
-            reverse("admin:payment_invoice_change", args=(invoice_0.pk,)),
+            reverse("admin:payment_proformainvoice_change", args=(invoice_0.pk,)),
         )
 
         self.assertEqual(
@@ -122,32 +124,33 @@ class InvoiceAdminTestCase(TestCase):
         )
         self.assertEqual(
             links[1].attrib["href"],
-            reverse("admin:payment_invoice_change", args=(invoice_1.pk,)),
+            reverse("admin:payment_proformainvoice_change", args=(invoice_1.pk,)),
         )
 
-    def test_admin_invoice_display_related_transactions(self):
+    def test_admin_proforma_invoice_display_related_transactions(self):
         """
-        Invoice admin view should display transactions related to the invoice instance.
+        ProformaInvoice admin view should display transactions related to
+        the pro forma invoice instance.
         """
         # - Login as admin
         user = factories.UserFactory(is_staff=True, is_superuser=True)
         self.client.login(username=user.username, password="password")
 
-        # - Create an order with a related invoice and transaction
+        # - Create an order with a related pro forma invoice and transaction
         order = factories.OrderFactory()
-        invoice = InvoiceFactory(order=order, total=order.total)
-        transaction = TransactionFactory(invoice=invoice, total=invoice.total)
+        invoice = ProformaInvoiceFactory(order=order, total=order.total)
+        transaction = TransactionFactory(proforma_invoice=invoice, total=invoice.total)
 
-        # - Now go to the invoice admin change view
+        # - Now go to the pro forma invoice admin change view
         response = self.client.get(
-            reverse("admin:payment_invoice_change", args=(invoice.pk,)),
+            reverse("admin:payment_proformainvoice_change", args=(invoice.pk,)),
         )
 
-        # - Invoice are ordered by creation date
+        # - ProformaInvoice are ordered by creation date
 
         self.assertEqual(response.status_code, 200)
 
-        # - Check there are links to go to invoice children admin change view
+        # - Check there are links to go to pro forma invoice children admin change view
         html_parser = lxml.html.HTMLParser(encoding="utf-8")
         html = lxml.html.fromstring(response.content, parser=html_parser)
         invoice_children_field = html.cssselect(".field-transactions")[0]